# api/openapi.yaml
openapi: 3.1.0
info:
  title: Envy API
  version: 0.7.0
  description: >-
    Compose a shared staging baseline with selected prebuilt workload overrides.
    This development slice provides immutable catalog registration, up to three approved component overrides,
    durable asynchronous lifecycle, and request routing isolation.
    Authentication depends on installation mode: Envy/machine bearer credentials,
    browser session cookies (password/Google), trusted proxy identity, or explicit dev/none.
    Browser mutations require same-origin Origin and X-CSRF-Token from /auth/config.
    OAuth discovery and MCP authorization are documented in docs/authentication-and-activity.md.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:8081
security:
  - bearerAuth: []
  - browserSession: []
  - localBrowserSession: []
...
paths:
...
  /v1/projects/{project}/frontend-bindings/{frontend}/{revision}:
    parameters: &id001
    - $ref: '#/components/parameters/Project'
    - name: frontend
      in: path
      required: true
      schema:
        type: string
    - name: revision
      in: path
      required: true
      schema:
        type: string
        pattern: ^(?:[a-f0-9]{40}|[a-f0-9]{64})$
    put:
      operationId: bindFrontend
      tags:
      - Frontend bindings
      summary: Bind an exact frontend commit immutably to a composition
      responses:
        '200':
          description: Current binding observation; repeated identical writes are successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendBindingView'
        default:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindFrontendRequest'
    get:
      operationId: getFrontendBinding
      tags:
      - Frontend bindings
      summary: Inspect binding, including expired associations
      responses:
        '200':
          description: Current binding observation; repeated identical writes are successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendBindingView'
        default:
          $ref: '#/components/responses/Error'
  /v1/projects/{project}/frontend-bindings/{frontend}/{revision}/resolve:
    parameters: *id001
    get:
      operationId: resolveFrontend
      tags:
      - Frontend bindings
      summary: Resolve ready API URL; 409 when unready, 410 after expiry/deletion; no
        fallback
      responses:
        '200':
          description: Current binding observation; repeated identical writes are successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendResolution'
        default:
          $ref: '#/components/responses/Error'
  /v1/projects/{project}/frontend-bindings/{frontend}/{revision}/deployment:
    parameters: *id001
    post:
      operationId: publishFrontend
      tags:
      - Frontend bindings
      summary: Record caller-reported URL with optimistic version checking; does not
        deploy hosting
      responses:
        '200':
          description: Current binding observation; repeated identical writes are successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendBindingView'
        default:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishFrontendRequest'
...
  /v1/compositions/{id}/frontend-bindings:
    parameters:
    - $ref: '#/components/parameters/CompositionID'
    - $ref: '#/components/parameters/After'
    - $ref: '#/components/parameters/Limit'
    get:
      operationId: listFrontendBindings
      tags:
      - Frontend bindings
      summary: List a composition’s exact frontend revisions
      responses:
        '200':
          description: Current binding observation; repeated identical writes are successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendBindingsPage'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/projects:
    post:
      operationId: registerProject
      tags: [Catalog]
      summary: Register an immutable project
      description: Duplicate IDs or routing host claims return 409. Baseline registration validates existing sidecar workloads, ingress normalization and routing ownership before persistence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: listProjects
      tags: [Catalog]
      summary: List registered projects
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/After'
      responses:
        '200':
          description: Bounded project list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Error'
  /v1/projects/{project}/components:
    post:
      operationId: registerComponent
      tags: [Catalog]
      summary: Register an immutable component
      description: Duplicate IDs or routing host claims return 409. Baseline registration validates existing sidecar workloads, ingress normalization and routing ownership before persistence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentRegistration'
      responses:
        '201':
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
        default:
          $ref: '#/components/responses/Error'
    parameters:
      - $ref: '#/components/parameters/Project'
    get:
      operationId: listComponents
      tags: [Catalog]
      summary: List a project's component profiles
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/After'
      responses:
        '200':
          description: Bounded component list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Component'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/projects/{project}/baselines:
    post:
      operationId: registerBaseline
      tags: [Catalog]
      summary: Register an immutable baseline
      description: Duplicate IDs or routing host claims return 409. Baseline registration validates existing sidecar workloads, ingress normalization and routing ownership before persistence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaselineRegistration'
      responses:
        '201':
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Baseline'
        default:
          $ref: '#/components/responses/Error'
    parameters:
      - $ref: '#/components/parameters/Project'
    get:
      operationId: listBaselines
      tags: [Catalog]
      summary: List a project's registered baselines
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/After'
      responses:
        '200':
          description: Bounded baseline list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Baseline'
        default:
          $ref: '#/components/responses/Error'
  /v1/compositions:
    post:
      operationId: createComposition
      tags: [Compositions]
      summary: Accept a composition and persist its create operation
      description: >-
        Provider work is asynchronous. A URL may be allocated immediately, but
        it is not ready until the expected request chain and baseline are verified.
        Same-key same-request retries return the original composition.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Optional retry key; different content with the same key returns 409.
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: '^[ -~]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComposition'
            example:
              project: demo
              baseline: staging
              name: service-b-v2
              overrides:
                service-b:
                  image: envy/service-b:v2
              ttl: 8h
      responses:
        '202':
          description: Accepted, or replay of the original accepted request
          headers:
            Location:
              description: Relative composition polling location
              schema:
                type: string
                example: /v1/compositions/01999999-0000-7000-8000-000000000001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Composition'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: listCompositions
      tags: [Compositions]
      summary: List durable compositions, including retained tombstones
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/After'
        - name: project
          in: query
          description: Optional project scope filter
          schema:
            type: string
      responses:
        '200':
          description: Bounded composition list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Composition'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-envy_session
      description: HTTPS browser session. Mutations additionally require Origin and X-CSRF-Token.
    localBrowserSession:
      type: apiKey
      in: cookie
      name: envy_session
      description: Loopback HTTP browser session for authenticated local testing.
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Project:
      name: project
      in: path
      required: true
      schema:
        type: string
        example: demo
    CompositionID:
      name: id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum returned items
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    After:
      name: after
      in: query
      description: Last ID from the preceding page's next_cursor
      schema:
        type: string
  responses:
...
    Error:
      description: Structured error; retryable indicates a potentially transient failure
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                $ref: '#/components/schemas/Error'
  schemas:
...
    BindFrontendRequest:
      type: object
      additionalProperties: false
      required:
      - composition
      - repository
      properties:
        composition:
          type: string
        repository:
          type: string
          format: uri
          description: HTTPS source repository; no credentials, query or fragment
    PublishFrontendRequest:
      type: object
      additionalProperties: false
      required:
      - expected_version
      - url
      properties:
        expected_version:
          type: integer
          format: int64
          minimum: 1
        url:
          type: string
          format: uri
          description: HTTPS; HTTP only for loopback. No credentials, query or fragment.
...
    FrontendBinding:
      type: object
      additionalProperties: false
      required:
      - project
      - frontend
      - revision
      - composition
      - repository
      - version
      - created_at
      - updated_at
      properties:
        project:
          type: string
        frontend:
          type: string
        revision:
          type: string
        composition:
          type: string
        repository:
          type: string
        version:
          type: integer
          format: int64
          minimum: 1
        url:
          type: string
        check:
          $ref: '#/components/schemas/FrontendCheck'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FrontendBindingView:
      type: object
      additionalProperties: false
      required:
      - binding
      - composition_phase
      - composition_generation
      - expires_at
      - ready
      - verification_level
      - check_state
      properties:
        binding:
          $ref: '#/components/schemas/FrontendBinding'
        composition_phase:
          type: string
        composition_generation:
          type: integer
          format: int64
          minimum: 1
        expires_at:
          type: string
          format: date-time
        ready:
          type: boolean
        verification_level:
          type: string
        check_state:
          type: string
          enum:
          - not_reported
          - current
          - stale
    FrontendResolution:
      type: object
      additionalProperties: false
      required:
      - project
      - frontend
      - revision
      - composition
      - composition_generation
      - binding_version
      - api_url
      - expires_at
      - verification_level
      properties:
        project:
          type: string
        frontend:
          type: string
        revision:
          type: string
        composition:
          type: string
        composition_generation:
          type: integer
          format: int64
          minimum: 1
        binding_version:
          type: integer
          format: int64
          minimum: 1
        api_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
        verification_level:
          type: string
    FrontendBindingsPage:
      type: object
      additionalProperties: false
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FrontendBindingView'
        next_cursor:
          type: string
...
    Page:
      type: object
      properties:
        next_cursor:
          type: string
          description: Last returned ID when another page exists; absent or empty on the final page
    Project:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          example: demo
        name:
          type: string
    ComponentRegistration:
      type: object
      additionalProperties: false
      required: [id, protocol, port, profile]
      allOf:
        - if:
            properties:
              profile: {const: http-small}
          then:
            required: [health_path, readiness_path]
            properties:
              port: {minimum: 1024}
          else:
            properties:
              health_path: {maxLength: 0}
              readiness_path: {maxLength: 0}
              env: {maxProperties: 0}
              image_pull_secrets: {maxItems: 0}
      properties:
        image_pull_secrets:
          type: array
          maxItems: 8
          uniqueItems: true
          items:
            type: string
            maxLength: 63
            pattern: '^[a-z0-9]([a-z0-9-]*[a-z0-9])?$'
          description: Operator-approved Secret names supplied externally in each composition namespace. No credential copying.
        profile:
          type: string
          enum: [http-small, deployment]
          description: Fixed legacy profile or opt-in deployment-derived configuration requiring preview approval
        readiness_path:
          type: string
        env:
          type: object
          maxProperties: 32
          additionalProperties:
            type: string
          description: Approved literal environment, at most 8 KiB; catalog-visible, no credentials or reserved ENVY_/POD_UID keys
        id:
          type: string
          example: service-b
        project:
          type: string
          example: demo
        protocol:
          type: string
          const: http
        port:
          type: integer
          minimum: 1
          maximum: 65535
        health_path:
          type: string
        overridable:
          type: boolean
        repository:
          type: string
          description: Optional source provenance, not a composition-wide branch identity
    Component:
      allOf:
        - $ref: '#/components/schemas/ComponentRegistration'
        - type: object
          required: [project]
    BaselineRegistration:
      type: object
      additionalProperties: false
      required: [id, revision, endpoint, components, routing, verification]
      properties:
        pubsub:
          type: object
          maxProperties: 20
          additionalProperties:
            $ref: '#/components/schemas/PubSubTopic'
        routing:
          type: object
          additionalProperties: false
          required: [namespace, gateway, entry_component]
          properties:
            namespace: {type: string}
            gateway: {type: string}
            entry_component: {type: string}
            gateway_namespace: {type: string}
            gateway_section_name: {type: string}
        verification:
          type: object
          additionalProperties: false
          required: [kind]
          oneOf:
            - properties:
                kind: {const: envy-chain}
              required: [chain]
              not:
                anyOf:
                  - required: [path]
                  - required: [expected_status]
            - properties:
                kind: {const: http}
              not:
                required: [chain]
          properties:
            kind: {type: string, enum: [envy-chain, http]}
            path: {type: string, default: /, description: Absolute HTTP probe path; only for http verification}
            expected_status: {type: integer, minimum: 200, maximum: 299, default: 200}
            chain:
              type: array
              minItems: 1
              maxItems: 20
              uniqueItems: true
              items: {type: string}
        id:
          type: string
          example: staging
        project:
          type: string
          example: demo
        revision:
          type: string
          description: Registered binding revision; inherited deployments remain live
        endpoint:
          type: string
          format: uri
          example: http://baseline.envy.localhost:8080
        components:
          type: object
          additionalProperties:
            type: object
            required: [service_host, port, image]
            properties:
              service_host:
                type: string
              port:
                type: integer
              image:
                type: string
    Baseline:
      allOf:
        - $ref: '#/components/schemas/BaselineRegistration'
        - type: object
          required: [project]
...
    ResolvedComponentOverride:
      type: object
      additionalProperties: false
      required: [image]
      properties:
        image:
          type: string
        build_id:
          type: string
        source:
          $ref: '#/components/schemas/PublishedBuild'
    CreateComposition:
      type: object
      additionalProperties: false
      required: [project, baseline, name, overrides]
      properties:
        message_isolation:
          type: boolean
          default: false
          description: Composition-wide Pub/Sub isolation, immutable for its lifetime. Consumer deployment is optional.
        expected_preview_revisions:
          type: object
          description: Optional approved profile guards for selected overrides; updates check captured revisions.
          additionalProperties: {type: integer, minimum: 1}
        expected_baseline_revision:
          type: string
          description: Optional binding revision guard. A mismatch returns 409 before accepting creation; inherited deployments remain live.
        project:
          type: string
        baseline:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 128
        overrides:
          type: object
          maxProperties: 3
          description: Zero to three approved overrides. An empty object creates a preview that inherits the complete baseline.
          additionalProperties:
            $ref: '#/components/schemas/ComponentOverride'
        ttl:
          type: string
          default: 8h
          description: Positive Go duration no greater than the configured maximum, default 24h
        resources:
          type: object
          maxProperties: 0
          description: Only an empty object is accepted; baseline resources are inherited
...
    Composition:
      allOf:
        - $ref: '#/components/schemas/CompositionStatus'
        - type: object
          required:
            - project
            - baseline
            - baseline_revision
            - name
            - overrides
            - expires_at
            - created_at
            - updated_at
            - components
            - endpoints
          properties:
            message_isolation:
              type: boolean
              default: false
            message_subscriptions:
              type: array
              items:
                $ref: '#/components/schemas/MessageSubscription'
            preview_profiles:
              type: object
              additionalProperties:
                type: object
                required: [revision, source]
                properties:
                  revision: {type: integer}
                  source: {$ref: '#/components/schemas/PreviewSource'}
            project:
              type: string
            baseline:
              type: string
            baseline_revision:
              type: string
              description: Binding revision; does not freeze inherited deployments
            name:
              type: string
            overrides:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/ResolvedComponentOverride'
            expires_at:
              type: string
              format: date-time
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            components:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/ComponentObservation'
            endpoints:
              $ref: '#/components/schemas/Endpoints'
...